apparently fixed tl2 as well

+++ stamp-0.9.10/kmeans/normal.c
@@ -110,7 +110,7 @@ typedef struct args {
 
 #if !defined(STO)
 float global_delta;
-float global_i;
+long global_i;
 #else
 Single<float> global_delta;
 Single<long> global_i; /* index into task queue */
